requirements.txt
# AI Industry Signals - Complete Requirements
# Install with: pip install -r requirements.txt

# ============================================
# Core ETL & Text Processing
# ============================================
PyMuPDF==1.23.8          # PDF extraction
pdfplumber==0.10.3       # PDF fallback extraction
tiktoken==0.5.2          # Token counting for chunking
beautifulsoup4==4.12.2   # HTML parsing
lxml==4.9.3              # XML/HTML parsing
requests==2.31.0         # HTTP requests for web scraping
python-dotenv==1.0.0     # Environment variables

# Document processing
python-docx==1.1.0       # Word document processing (NOT docx)
openpyxl==3.1.2          # Excel file processing
PyPDF2==3.0.1            # Additional PDF processing
chardet==5.2.0           # Character encoding detection
unidecode==1.3.8         # Unicode handling

# Audio processing
faster-whisper==1.0.0    # Optimized Whisper (4x faster, recommended)
openai-whisper==20231117 # Audio transcription (fallback)
soundfile==0.12.1        # Audio file handling
librosa==0.10.1          # Audio analysis

# ============================================
# Database - Relational
# ============================================
duckdb==0.9.2            # Embedded SQL database (primary)
psycopg2-binary==2.9.9   # PostgreSQL driver (if using PostgreSQL)

# ============================================
# Database - Vector & Embeddings
# ============================================
faiss-cpu==1.7.4         # Vector similarity search (CPU version for Jetson)
# For GPU support on Jetson, install via: pip install faiss-gpu
sentence-transformers==2.2.2  # Text embeddings
huggingface-hub==0.19.4   # HuggingFace hub (compatible version)
torch==2.1.0             # PyTorch (required for sentence-transformers)
torchvision==0.16.0      # PyTorch vision (dependency)
transformers==4.35.2     # Hugging Face transformers
numpy==1.24.3            # Numerical computing
scipy==1.11.4            # Scientific computing

# CUDA support (Jetson)
# Note: On Jetson, install PyTorch with CUDA support:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

# ============================================
# LLM Integration
# ============================================
# For Ollama (local LLM)
# No package needed - uses HTTP API via requests

# Alternative LLM options (uncomment if needed):
# openai==1.3.7          # OpenAI API
# anthropic==0.7.1       # Claude API
# cohere==4.37           # Cohere API
# langchain==0.0.350     # LangChain framework (if needed later)
# langchain-community==0.0.1  # LangChain community

# ============================================
# API Server & Web Framework
# ============================================
fastapi==0.104.1         # API framework
uvicorn==0.24.0          # ASGI server
pydantic==2.5.0          # Data validation
python-multipart==0.0.6  # Form parsing

# ============================================
# Utilities & Tools
# ============================================
python-dateutil==2.8.2   # Date parsing
pytz==2023.3             # Timezone handling
tqdm==4.66.1             # Progress bars
colorama==0.4.6          # Colored terminal output

# ============================================
# Logging & Monitoring
# ============================================
loguru==0.7.2            # Enhanced logging (optional upgrade)

# ============================================
# Data Processing & Analysis
# ============================================
pandas==2.1.3            # Data manipulation (for analytics)
scikit-learn==1.3.2      # Machine learning (for clustering)

# ============================================
# Testing (Development only)
# ============================================
pytest==7.4.3            # Testing framework
pytest-asyncio==0.21.1   # Async testing
httpx==0.25.2            # Async HTTP client for testing

# ============================================
# Type Checking & Code Quality (Development only)
# ============================================
mypy==1.7.1              # Type checking
black==23.12.0           # Code formatting
flake8==6.1.0            # Linting
isort==5.13.2            # Import sorting
